backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlmodel==0.0.19
SQLAlchemy==2.0.48
pydantic==2.12.5
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose==3.3.0
bcrypt==4.0.1
email-validator==2.3.0
aiofiles==23.2.1

numpy==2.4.4
pandas==3.0.1
scikit-learn==1.8.0
joblib==1.5.3

tensorflow==2.21.0
keras==3.13.2

opencv-python==4.13.0.92
opencv-contrib-python==4.13.0.92
mediapipe==0.10.33

librosa==0.11.0
soundfile==0.13.1
webrtcvad==2.0.10
faster-whisper==1.2.1
pydub==0.25.1

requests==2.33.0